migrate project svarga page to app directory
