+++ constants.ts
@@ -1,8 +1,2 @@
 export const ENTRY_PER_PAGE = 30;
 export const PAGE_PER_FOLD = 5;
-
-export const BLOG_TAGS = {
-    GENERAL: 'general',
-    JAVASCript: 'javascript',
-    STATIC_SITES: 'static-sites'
-};
