+++ package.json
@@ -22,7 +22,7 @@
         "react-bootstrap": "^2.10.0",
         "react-dom": "^18.2.0",
         "react-markdown": "^9.0.1",
-        "react-redux": "^7.2.2",
+        "react-redux": "^9.1.0",
         "redux": "^4.0.5",
         "redux-devtools-extension": "^2.13.8",
         "redux-thunk": "^2.3.0",
